+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Planned: April 2025
 
 ### Corrected Bugs
 
+* Fixed: [Diff of top-level default values on startup stopped working in 7.3](https://github.com/clicon/clixon/issues/596)
 * Fixed: [cli_show_auto don't work](https://github.com/clicon/clixon/issues/595)
 * Fixed: [XPath * stopped working in 7.3](https://github.com/clicon/clixon/issues/594)
 * Fixed: [Templates with nc:operation "merge" causes bad diffs to be shows](https://github.com/clicon/clixon-controller/issues/187)